refactor(api): move controller to src/team
