+++ api/lib/domain/usecases/index.js
@@ -151,6 +151,7 @@ import * as dateUtils from '../../../src/shared/infrastructure/utils/date-utils.
 import { injectDependencies } from '../../../src/shared/infrastructure/utils/dependency-injection.js';
 import { importNamedExportsFromDirectory } from '../../../src/shared/infrastructure/utils/import-named-exports-from-directory.js';
 import * as emailRepository from '../../../src/shared/mail/infrastructure/repositories/email.repository.js';
+// Not used in lib
 import * as certificationCenterInvitationService from '../../../src/team/domain/services/certification-center-invitation-service.js';
 import { organizationInvitationService } from '../../../src/team/domain/services/organization-invitation.service.js';
 import * as certificationCenterInvitationRepository from '../../../src/team/infrastructure/repositories/certification-center-invitation-repository.js';
