+++ api/src/team/application/certification-center-invitation/certification-center-invitation.controller.js
@@ -1,4 +1,7 @@
-import { extractLocaleFromRequest } from '../../../shared/infrastructure/utils/request-response-utils.js';
+import {
+  extractLocaleFromRequest,
+  requestResponseUtils,
+} from '../../../shared/infrastructure/utils/request-response-utils.js';
 import { usecases } from '../../domain/usecases/index.js';
 import { certificationCenterInvitationSerializer } from '../../infrastructure/serializers/jsonapi/certification-center-invitation-serializer.js';
 
@@ -50,6 +53,18 @@ const sendInvitations = async function (request, h) {
   return h.response().code(204);
 };
 
+const resendCertificationCenterInvitation = async function (request, h) {
+  const certificationCenterInvitationId = request.params.certificationCenterInvitationId;
+  const locale = requestResponseUtils.extractLocaleFromRequest(request);
+
+  const certificationCenterInvitation = await usecases.resendCertificationCenterInvitation({
+    certificationCenterInvitationId,
+    locale,
+  });
+
+  return h.response(certificationCenterInvitationSerializer.serializeForAdmin(certificationCenterInvitation)).code(200);
+};
+
 const cancelCertificationCenterInvitation = async function (request, h) {
   const certificationCenterInvitationId = request.params.certificationCenterInvitationId;
   await usecases.cancelCertificationCenterInvitation({ certificationCenterInvitationId });
@@ -70,5 +85,6 @@ export const certificationCenterInvitationController = {
   cancelCertificationCenterInvitation,
   getCertificationCenterInvitation,
   findPendingInvitations,
+  resendCertificationCenterInvitation,
   sendInvitations,
 };
