+++ api/tests/team/unit/application/certification-center-invitation/certification-center-invitation.controller.test.js
@@ -1,6 +1,8 @@
+import { requestResponseUtils } from '../../../../../src/shared/infrastructure/utils/request-response-utils.js';
 import { certificationCenterInvitationController } from '../../../../../src/team/application/certification-center-invitation/certification-center-invitation.controller.js';
 import { usecases } from '../../../../../src/team/domain/usecases/index.js';
-import { expect, hFake, sinon } from '../../../../test-helper.js';
+import { certificationCenterInvitationSerializer } from '../../../../../src/team/infrastructure/serializers/jsonapi/certification-center-invitation-serializer.js';
+import { domainBuilder, expect, hFake, sinon } from '../../../../test-helper.js';
 
 describe('Unit | Team | Application | Controller | Certification-center-invitation', function () {
   describe('#acceptCertificationCenterInvitation', function () {
@@ -33,6 +35,56 @@ describe('Unit | Team | Application | Controller | Certification-center-invitati
         localeFromCookie: undefined,
       });
       expect(response.statusCode).to.equal(204);
+    });
+  });
+
+  describe('#resendCertificationCenterInvitation', function () {
+    it('calls the resend certification center invitation usecase and returns an certification center invitation', async function () {
+      // given
+      const certificationCenterInvitationId = 123;
+      const certificationCenterInvitation = domainBuilder.buildCertificationCenterInvitation({
+        id: certificationCenterInvitationId,
+      });
+      const locale = 'nl-BE';
+      const serializerResult = {
+        type: 'certification-center-invitation',
+        id: certificationCenterInvitation.id,
+        attributes: {
+          email: certificationCenterInvitation.email,
+          role: certificationCenterInvitation.role,
+          'updated-at': certificationCenterInvitation.updatedAt,
+        },
+      };
+
+      sinon.stub(requestResponseUtils, 'extractLocaleFromRequest').returns(locale);
+
+      sinon.stub(usecases, 'resendCertificationCenterInvitation');
+      usecases.resendCertificationCenterInvitation
+        .withArgs({
+          certificationCenterInvitationId,
+          locale,
+        })
+        .resolves(certificationCenterInvitation);
+
+      sinon.stub(certificationCenterInvitationSerializer, 'serializeForAdmin');
+      certificationCenterInvitationSerializer.serializeForAdmin
+        .withArgs(certificationCenterInvitation)
+        .returns(serializerResult);
+
+      // when
+      const response = await certificationCenterInvitationController.resendCertificationCenterInvitation(
+        {
+          auth: { credentials: { userId: 1 } },
+          params: { certificationCenterInvitationId },
+        },
+        hFake,
+      );
+
+      // then
+      expect(requestResponseUtils.extractLocaleFromRequest).to.have.been.called;
+      expect(usecases.resendCertificationCenterInvitation).to.have.been.called;
+      expect(response.statusCode).to.equal(200);
+      expect(response.source).to.deep.equal(serializerResult);
     });
   });
 });
